refactor: deprecate token manager types and implementations
