+++ packages/backend-common/src/index.ts
@@ -22,9 +22,9 @@
 
 export { legacyPlugin, makeLegacyPlugin } from './legacy';
 export type { LegacyCreateRouter } from './legacy';
+export { loadBackendConfig } from './config';
 export * from './auth';
 export * from './cache';
-export { loadBackendConfig } from './config';
 export * from './deprecated';
 export * from './database';
 export * from './discovery';
@@ -33,5 +33,4 @@ export * from './logging';
 export * from './middleware';
 export * from './reading';
 export * from './service';
-export * from './tokens';
 export * from './util';
